Melhor modularização para legibilidade - Identificação do bug no carregamento de dados via debug | Ps: Encontrar uma forma de corrigir o bug antes do próximo commit.
